Move `DataType` class to `Yiisoft\Db\Constant` namespace
