+++ tests/Provider/CommandProvider.php
@@ -7,7 +7,7 @@ namespace Yiisoft\Db\Tests\Provider;
 use ArrayIterator;
 use IteratorAggregate;
 use Traversable;
-use Yiisoft\Db\Command\DataType;
+use Yiisoft\Db\Constant\DataType;
 use Yiisoft\Db\Command\Param;
 use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Expression\Expression;
